refactor: split assignment statement parser
