+++ src/parser/stmt/assign.rs
@@ -1,479 +1,18 @@
-use crate::errors::CompileError;
-use crate::lexer::Token;
-use crate::parser::ast::{Expr, ExprKind, Stmt, StmtKind};
-use crate::parser::expr::{parse_assignment_value_expr, parse_expr};
-use crate::span::Span;
+mod compound;
+mod locals;
+mod postfix;
+mod simple;
 
-use super::params::parse_type_expr;
-use super::{expect_semicolon, expect_token};
-
-/// Handle statements starting with $variable: assignment, array ops, or post-increment.
-pub(super) fn parse_variable_stmt(
-    tokens: &[(Token, Span)],
-    pos: &mut usize,
-    span: Span,
-) -> Result<Stmt, CompileError> {
-    let name = match &tokens[*pos].0 {
-        Token::Variable(n) => n.clone(),
-        _ => unreachable!(),
-    };
-
-    if let Some(stmt) = try_parse_postfix_assignment(tokens, pos, span)? {
-        return Ok(stmt);
-    }
-
-    // Post-increment/decrement
-    if *pos + 1 < tokens.len() {
-        match &tokens[*pos + 1].0 {
-            Token::PlusPlus => {
-                *pos += 2;
-                expect_semicolon(tokens, pos)?;
-                let expr = Expr::new(ExprKind::PostIncrement(name), span);
-                return Ok(Stmt::new(StmtKind::ExprStmt(expr), span));
-            }
-            Token::MinusMinus => {
-                *pos += 2;
-                expect_semicolon(tokens, pos)?;
-                let expr = Expr::new(ExprKind::PostDecrement(name), span);
-                return Ok(Stmt::new(StmtKind::ExprStmt(expr), span));
-            }
-            _ => {}
-        }
-    }
-
-    if *pos + 1 < tokens.len()
-        && matches!(
-            tokens[*pos + 1].0,
-            Token::Arrow | Token::QuestionArrow | Token::LBracket
-        )
-    {
-        let expr = parse_expr(tokens, pos)?;
-        if *pos < tokens.len() && tokens[*pos].0 == Token::Assign {
-            *pos += 1;
-            let value = parse_assignment_value_expr(tokens, pos)?;
-            expect_semicolon(tokens, pos)?;
-            if let ExprKind::PropertyAccess { object, property } = expr.kind {
-                return Ok(Stmt::new(
-                    StmtKind::PropertyAssign {
-                        object,
-                        property,
-                        value,
-                    },
-                    span,
-                ));
-            }
-            return Err(CompileError::new(span, "Invalid assignment target"));
-        }
-        expect_semicolon(tokens, pos)?;
-        return Ok(Stmt::new(StmtKind::ExprStmt(expr), span));
-    }
-
-    // Closure call: $fn(args);
-    if *pos + 1 < tokens.len() && tokens[*pos + 1].0 == Token::LParen {
-        let expr = parse_expr(tokens, pos)?;
-        expect_semicolon(tokens, pos)?;
-        return Ok(Stmt::new(StmtKind::ExprStmt(expr), span));
-    }
-
-    // Regular or compound assignment
-    parse_assign(tokens, pos, span)
-}
-
-pub(super) fn parse_assign(
-    tokens: &[(Token, Span)],
-    pos: &mut usize,
-    span: Span,
-) -> Result<Stmt, CompileError> {
-    let name = match &tokens[*pos].0 {
-        Token::Variable(n) => n.clone(),
-        _ => unreachable!(),
-    };
-    *pos += 1;
-
-    if *pos >= tokens.len() {
-        return Err(CompileError::new(span, "Expected '=' after variable name"));
-    }
-
-    use crate::parser::ast::BinOp;
-    let compound_op = match &tokens[*pos].0 {
-        Token::PlusAssign => Some(BinOp::Add),
-        Token::MinusAssign => Some(BinOp::Sub),
-        Token::StarAssign => Some(BinOp::Mul),
-        Token::StarStarAssign => Some(BinOp::Pow),
-        Token::SlashAssign => Some(BinOp::Div),
-        Token::PercentAssign => Some(BinOp::Mod),
-        Token::DotAssign => Some(BinOp::Concat),
-        Token::AmpAssign => Some(BinOp::BitAnd),
-        Token::PipeAssign => Some(BinOp::BitOr),
-        Token::CaretAssign => Some(BinOp::BitXor),
-        Token::LessLessAssign => Some(BinOp::ShiftLeft),
-        Token::GreaterGreaterAssign => Some(BinOp::ShiftRight),
-        Token::Assign => None,
-        Token::QuestionQuestionAssign => {
-            *pos += 1;
-            let rhs = parse_assignment_value_expr(tokens, pos)?;
-            expect_semicolon(tokens, pos)?;
-            let value = Expr::new(
-                ExprKind::NullCoalesce {
-                    value: Box::new(Expr::new(ExprKind::Variable(name.clone()), span)),
-                    default: Box::new(rhs),
-                },
-                span,
-            );
-            return Ok(Stmt::new(StmtKind::Assign { name, value }, span));
-        }
-        _ => return Err(CompileError::new(span, "Expected '=' after variable name")),
-    };
-    *pos += 1;
-
-    let rhs = parse_assignment_value_expr(tokens, pos)?;
-    expect_semicolon(tokens, pos)?;
-
-    let value = if let Some(op) = compound_op {
-        Expr::new(
-            ExprKind::BinaryOp {
-                left: Box::new(Expr::new(ExprKind::Variable(name.clone()), span)),
-                op,
-                right: Box::new(rhs),
-            },
-            span,
-        )
-    } else {
-        rhs
-    };
-
-    Ok(Stmt::new(StmtKind::Assign { name, value }, span))
-}
-
-pub(super) fn try_parse_postfix_assignment(
-    tokens: &[(Token, Span)],
-    pos: &mut usize,
-    span: Span,
-) -> Result<Option<Stmt>, CompileError> {
-    let start = *pos;
-    let Some(assign_pos) = find_top_level_assign(tokens, start) else {
-        return Ok(None);
-    };
-    if assign_pos < start + 3 {
-        return Ok(None);
-    }
-
-    let lhs = &tokens[start..assign_pos];
-    let is_append =
-        lhs.len() >= 3 && lhs[lhs.len() - 2].0 == Token::LBracket && lhs[lhs.len() - 1].0 == Token::RBracket;
-    let contains_postfix = lhs
-        .iter()
-        .skip(1)
-        .any(|(token, _)| matches!(token, Token::Arrow | Token::QuestionArrow | Token::LBracket));
-    if !contains_postfix {
-        return Ok(None);
-    }
-
-    let mut lhs_pos = 0;
-    let lhs_expr_tokens = if is_append { &lhs[..lhs.len() - 2] } else { lhs };
-    let lhs_expr = parse_expr(lhs_expr_tokens, &mut lhs_pos)?;
-    if lhs_pos != lhs_expr_tokens.len() {
-        return Err(CompileError::new(span, "Invalid assignment target"));
-    }
-
-    *pos = assign_pos + 1;
-    let value = parse_assignment_value_expr(tokens, pos)?;
-    expect_semicolon(tokens, pos)?;
-
-    let stmt = match lhs_expr.kind {
-        ExprKind::Variable(array) if is_append => StmtKind::ArrayPush { array, value },
-        ExprKind::PropertyAccess { object, property } if is_append => StmtKind::PropertyArrayPush {
-            object,
-            property,
-            value,
-        },
-        ExprKind::ArrayAccess { array, index } => {
-            match array.kind {
-                ExprKind::Variable(array) => StmtKind::ArrayAssign {
-                    array,
-                    index: *index,
-                    value,
-                },
-                ExprKind::PropertyAccess { object, property } => StmtKind::PropertyArrayAssign {
-                    object,
-                    property,
-                    index: *index,
-                    value,
-                },
-                _ => return Err(CompileError::new(span, "Invalid assignment target")),
-            }
-        }
-        ExprKind::PropertyAccess { object, property } => StmtKind::PropertyAssign {
-            object,
-            property,
-            value,
-        },
-        _ => return Err(CompileError::new(span, "Invalid assignment target")),
-    };
-
-    Ok(Some(Stmt::new(stmt, span)))
-}
-
-pub(super) fn try_parse_scoped_property_assignment(
-    tokens: &[(Token, Span)],
-    pos: &mut usize,
-    span: Span,
-) -> Result<Option<Stmt>, CompileError> {
-    let start = *pos;
-    let Some(assign_pos) = find_top_level_assign(tokens, start) else {
-        return Ok(None);
-    };
-    if assign_pos < start + 3 {
-        return Ok(None);
-    }
-
-    let lhs = &tokens[start..assign_pos];
-    let is_append =
-        lhs.len() >= 3 && lhs[lhs.len() - 2].0 == Token::LBracket && lhs[lhs.len() - 1].0 == Token::RBracket;
-    let mut lhs_pos = 0;
-    let lhs_expr_tokens = if is_append { &lhs[..lhs.len() - 2] } else { lhs };
-    let lhs_expr = parse_expr(lhs_expr_tokens, &mut lhs_pos)?;
-    if lhs_pos != lhs_expr_tokens.len() {
-        return Err(CompileError::new(span, "Invalid assignment target"));
-    }
-
-    *pos = assign_pos + 1;
-    let value = parse_assignment_value_expr(tokens, pos)?;
-    expect_semicolon(tokens, pos)?;
-
-    let stmt = match lhs_expr.kind {
-        ExprKind::StaticPropertyAccess { receiver, property } if is_append => {
-            StmtKind::StaticPropertyArrayPush {
-                receiver,
-                property,
-                value,
-            }
-        }
-        ExprKind::ArrayAccess { array, index } => match array.kind {
-            ExprKind::StaticPropertyAccess { receiver, property } => {
-                StmtKind::StaticPropertyArrayAssign {
-                    receiver,
-                    property,
-                    index: *index,
-                    value,
-                }
-            }
-            _ => return Err(CompileError::new(span, "Invalid assignment target")),
-        },
-        ExprKind::StaticPropertyAccess { receiver, property } => StmtKind::StaticPropertyAssign {
-            receiver,
-            property,
-            value,
-        },
-        _ => return Err(CompileError::new(span, "Invalid assignment target")),
-    };
-
-    Ok(Some(Stmt::new(stmt, span)))
-}
-
-fn find_top_level_assign(tokens: &[(Token, Span)], start: usize) -> Option<usize> {
-    let mut paren_depth = 0usize;
-    let mut bracket_depth = 0usize;
-    let mut brace_depth = 0usize;
-    let mut pos = start;
-
-    while pos < tokens.len() {
-        match tokens[pos].0 {
-            Token::LParen => paren_depth += 1,
-            Token::RParen => paren_depth = paren_depth.saturating_sub(1),
-            Token::LBracket => bracket_depth += 1,
-            Token::RBracket => bracket_depth = bracket_depth.saturating_sub(1),
-            Token::LBrace => brace_depth += 1,
-            Token::RBrace => brace_depth = brace_depth.saturating_sub(1),
-            Token::Assign if paren_depth == 0 && bracket_depth == 0 && brace_depth == 0 => {
-                return Some(pos);
-            }
-            Token::Semicolon if paren_depth == 0 && bracket_depth == 0 && brace_depth == 0 => {
-                return None;
-            }
-            _ => {}
-        }
-        pos += 1;
-    }
-
-    None
-}
-
-/// Handle ++$var; or --$var; as standalone statements.
-pub(super) fn parse_incdec_stmt(
-    tokens: &[(Token, Span)],
-    pos: &mut usize,
-    span: Span,
-) -> Result<Stmt, CompileError> {
-    let is_increment = tokens[*pos].0 == Token::PlusPlus;
-    *pos += 1;
-
-    let name = match tokens.get(*pos).map(|(t, _)| t) {
-        Some(Token::Variable(n)) => n.clone(),
-        _ => {
-            let op = if is_increment { "++" } else { "--" };
-            return Err(CompileError::new(
-                span,
-                &format!("Expected variable after '{}'", op),
-            ));
-        }
-    };
-    *pos += 1;
-    expect_semicolon(tokens, pos)?;
-
-    let kind = if is_increment {
-        ExprKind::PreIncrement(name)
-    } else {
-        ExprKind::PreDecrement(name)
-    };
-    let expr = Expr::new(kind, span);
-    Ok(Stmt::new(StmtKind::ExprStmt(expr), span))
-}
-
-pub(super) fn parse_list_unpack(
-    tokens: &[(Token, Span)],
-    pos: &mut usize,
-    span: Span,
-) -> Result<Stmt, CompileError> {
-    *pos += 1; // consume '['
-
-    let mut vars = Vec::new();
-    while *pos < tokens.len() && tokens[*pos].0 != Token::RBracket {
-        if !vars.is_empty() {
-            if tokens[*pos].0 != Token::Comma {
-                return Err(CompileError::new(
-                    tokens[*pos].1,
-                    "Expected ',' between list variables",
-                ));
-            }
-            *pos += 1;
-        }
-        match tokens.get(*pos).map(|(t, _)| t) {
-            Some(Token::Variable(n)) => {
-                vars.push(n.clone());
-                *pos += 1;
-            }
-            _ => {
-                return Err(CompileError::new(
-                    span,
-                    "Expected variable in list unpacking",
-                ))
-            }
-        }
-    }
-
-    if *pos >= tokens.len() || tokens[*pos].0 != Token::RBracket {
-        return Err(CompileError::new(span, "Expected ']' after list variables"));
-    }
-    *pos += 1;
-
-    expect_token(
-        tokens,
-        pos,
-        &Token::Assign,
-        "Expected '=' after list pattern",
-    )?;
-
-    let value = parse_assignment_value_expr(tokens, pos)?;
-    expect_semicolon(tokens, pos)?;
-
-    Ok(Stmt::new(StmtKind::ListUnpack { vars, value }, span))
-}
-
-pub(super) fn parse_global(
-    tokens: &[(Token, Span)],
-    pos: &mut usize,
-    span: Span,
-) -> Result<Stmt, CompileError> {
-    *pos += 1; // consume 'global'
-
-    let mut vars = Vec::new();
-    loop {
-        match tokens.get(*pos).map(|(t, _)| t) {
-            Some(Token::Variable(n)) => {
-                vars.push(n.clone());
-                *pos += 1;
-            }
-            _ => return Err(CompileError::new(span, "Expected variable after 'global'")),
-        }
-        if *pos < tokens.len() && tokens[*pos].0 == Token::Comma {
-            *pos += 1;
-        } else {
-            break;
-        }
-    }
-
-    expect_semicolon(tokens, pos)?;
-    Ok(Stmt::new(StmtKind::Global { vars }, span))
-}
-
-pub(super) fn parse_static_var(
-    tokens: &[(Token, Span)],
-    pos: &mut usize,
-    span: Span,
-) -> Result<Stmt, CompileError> {
-    *pos += 1; // consume 'static'
-
-    let name = match tokens.get(*pos).map(|(t, _)| t) {
-        Some(Token::Variable(n)) => n.clone(),
-        _ => return Err(CompileError::new(span, "Expected variable after 'static'")),
-    };
-    *pos += 1;
-
-    expect_token(
-        tokens,
-        pos,
-        &Token::Assign,
-        "Expected '=' after static variable",
-    )?;
-
-    let init = parse_assignment_value_expr(tokens, pos)?;
-    expect_semicolon(tokens, pos)?;
-
-    Ok(Stmt::new(StmtKind::StaticVar { name, init }, span))
-}
-
-pub(super) fn looks_like_typed_assign(tokens: &[(Token, Span)], pos: usize) -> bool {
-    let mut probe = pos;
-    match parse_type_expr(tokens, &mut probe, tokens[pos].1) {
-        Ok(_) => matches!(tokens.get(probe).map(|(t, _)| t), Some(Token::Variable(_))),
-        Err(_) => false,
-    }
-}
-
-pub(super) fn parse_typed_assign(
-    tokens: &[(Token, Span)],
-    pos: &mut usize,
-    span: Span,
-) -> Result<Stmt, CompileError> {
-    let type_expr = parse_type_expr(tokens, pos, span)?;
-    let name = match tokens.get(*pos).map(|(t, _)| t) {
-        Some(Token::Variable(name)) => {
-            let name = name.clone();
-            *pos += 1;
-            name
-        }
-        _ => {
-            return Err(CompileError::new(
-                span,
-                "Expected variable after type annotation",
-            ))
-        }
-    };
-    expect_token(
-        tokens,
-        pos,
-        &Token::Assign,
-        "Expected '=' after typed variable",
-    )?;
-    let value = parse_assignment_value_expr(tokens, pos)?;
-    expect_semicolon(tokens, pos)?;
-    Ok(Stmt::new(
-        StmtKind::TypedAssign {
-            type_expr,
-            name,
-            value,
-        },
-        span,
-    ))
-}
+pub(super) use locals::{
+    looks_like_typed_assign,
+    parse_global,
+    parse_incdec_stmt,
+    parse_list_unpack,
+    parse_static_var,
+    parse_typed_assign,
+};
+pub(super) use postfix::{
+    try_parse_postfix_assignment,
+    try_parse_scoped_property_assignment,
+};
+pub(super) use simple::parse_variable_stmt;
